Tutorial update
disabled player control when player is not supposed to move in the tutorial section. The tutorial should be more robust to spam input now. Please play with it and let me know what other ways you find that can spam the tutorial. I know it's not perfect yet. It's not as easy as I thought to make it fully robust